Split gallery.tsx into reusable modules
